refactor(flui-painting): Mythos Step 6 -- split text_layout.rs + flatten cfg
Split the 1,243-LOC `crates/flui-painting/src/text_layout.rs` god
module into a concern-based `text_layout/` directory. Five files,
each focused on one concern; total LOC drops to ~1,161 (net -82 from
removing duplicate type definitions between `inner` and `fallback`
modules + collapsed boilerplate).

Files created:

| File                                  | LOC | Concern                                                             |
|---------------------------------------|-----|---------------------------------------------------------------------|
| `text_layout/mod.rs`                  | 324 | Shared types (`TextLayoutResult` + `LineInfo`) + cfg-gated re-exports + inline tests |
| `text_layout/detect.rs`               |  65 | `#[cfg(feature = "text")]` RTL/LTR detection + Unicode codepoint helpers |
| `text_layout/layout.rs`               | 411 | `#[cfg(feature = "text")]` `FONT_SYSTEM` static + `TextLayout` struct + cursor/hit-test/line-metric methods |
| `text_layout/measure.rs`              | 123 | `#[cfg(feature = "text")]` `measure_text` + `measure_inline_span` + `style_to_attrs` helper |
| `text_layout/fallback.rs`             | 238 | `#[cfg(not(feature = "text"))]` stub `TextLayout` + stub `detect_text_direction` + stub `measure_*` |

Two structural improvements beyond the file split:

1. **Flatten the `mod inner` indirection.** The old file wrapped the
   real cosmic-text impl inside `#[cfg(feature = "text")] mod inner
   { ... }`. The new shape gates each sub-module file with `#[cfg]`
   directly; no inner module wrapping.

2. **Shared types live at the module root.** `TextLayoutResult` and
   `LineInfo` were duplicated between `inner` (cosmic-text impl) and
   `fallback` (stub impl) -- identical struct shapes. Moved to
   `text_layout/mod.rs` without cfg gating. Both implementations now
   import them.

Public API surface unchanged. External callers use the same
`flui_painting::{TextLayout, TextLayoutResult, LineInfo,
detect_text_direction, measure_text, measure_inline_span}` paths.
The `--no-default-features` build still produces a functional
fallback impl with estimated metrics.

The 19 inline tests from the old `text_layout.rs` are preserved in
`text_layout/mod.rs` for this step (gated `#[cfg(all(test, feature =
"text"))]`). U8 of the chain extracts them to
`crates/flui-painting/tests/text_layout_detect.rs` etc.

Verification:
- `cargo check -p flui-painting` clean.
- `cargo test -p flui-painting --lib` 64 passing (matches
  pre-refactor count).
- `cargo test -p flui-painting --tests` 9 passing.
- `cargo check --workspace` clean.
- `cargo build --no-default-features -p flui-painting` clean
  (fallback compiles standalone).
- `bash scripts/port-check.sh` 6 triggers ok.

Net unsafe delta: 0 (crate is #[forbid(unsafe_code)]; no unsafe
introduced; none removed).
Net LOC delta in production code: -82 LOC across the split files vs.
the original god module.

Mythos chain step 6 of 14. Next: U7 splits text_painter.rs (990 LOC)
into a `text_painter/` directory.
