--- crates/flui-painting/src/text_layout.rs
@@ -1,1243 +1,0 @@
-//! Text layout engine using cosmic-text.
-//!
-//! This module provides text shaping, measurement, and layout capabilities
-//! using the cosmic-text library.
-
-#[cfg(feature = "text")]
-mod inner {
-    use std::sync::OnceLock;
-
-    use cosmic_text::{Attrs, Buffer, Cursor, Family, FontSystem, Metrics, Shaping, Style, Weight};
-    use flui_types::{
-        geometry::{Offset, Pixels, Rect, Size, px},
-        typography::{
-            FontStyle, FontWeight, LineMetrics, TextAffinity, TextBox, TextDirection, TextPosition,
-            TextRange, TextStyle,
-        },
-    };
-    use parking_lot::Mutex;
-
-    /// Global font system instance.
-    ///
-    /// cosmic-text requires a FontSystem for font discovery and shaping.
-    /// We use a global instance with interior mutability for convenience.
-    static FONT_SYSTEM: OnceLock<Mutex<FontSystem>> = OnceLock::new();
-
-    /// Gets or initializes the global font system.
-    fn font_system() -> &'static Mutex<FontSystem> {
-        FONT_SYSTEM.get_or_init(|| {
-            tracing::debug!("Initializing global FontSystem");
-            Mutex::new(FontSystem::new())
-        })
-    }
-
-    /// Detects the text direction from the content.
-    ///
-    /// Uses the first strong directional character to determine direction.
-    /// Returns `None` if no strong directional character is found.
-    ///
-    /// # Examples
-    ///
-    /// ```rust,ignore
-    /// use flui_painting::text_layout::detect_text_direction;
-    /// use flui_types::typography::TextDirection;
-    ///
-    /// assert_eq!(detect_text_direction("Hello"), Some(TextDirection::Ltr));
-    /// assert_eq!(detect_text_direction("مرحبا"), Some(TextDirection::Rtl));
-    /// assert_eq!(detect_text_direction("123"), None); // Neutral
-    /// ```
-    pub fn detect_text_direction(text: &str) -> Option<TextDirection> {
-        for ch in text.chars() {
-            // Check for RTL characters (Arabic, Hebrew, etc.)
-            if is_rtl_char(ch) {
-                return Some(TextDirection::Rtl);
-            }
-            // Check for LTR characters (Latin, Greek, Cyrillic, etc.)
-            if is_ltr_char(ch) {
-                return Some(TextDirection::Ltr);
-            }
-            // Skip neutral characters (numbers, spaces, punctuation)
-        }
-        None
-    }
-
-    /// Checks if a character is a strong RTL character.
-    fn is_rtl_char(ch: char) -> bool {
-        matches!(ch,
-            // Arabic
-            '\u{0600}'..='\u{06FF}' |
-            '\u{0750}'..='\u{077F}' |
-            '\u{08A0}'..='\u{08FF}' |
-            '\u{FB50}'..='\u{FDFF}' |
-            '\u{FE70}'..='\u{FEFF}' |
-            // Hebrew
-            '\u{0590}'..='\u{05FF}' |
-            '\u{FB1D}'..='\u{FB4F}' |
-            // Syriac
-            '\u{0700}'..='\u{074F}' |
-            // Thaana
-            '\u{0780}'..='\u{07BF}' |
-            // N'Ko
-            '\u{07C0}'..='\u{07FF}'
-        )
-    }
-
-    /// Checks if a character is a strong LTR character.
-    fn is_ltr_char(ch: char) -> bool {
-        matches!(ch,
-            // Basic Latin letters
-            'A'..='Z' | 'a'..='z' |
-            // Latin Extended
-            '\u{00C0}'..='\u{024F}' |
-            // Greek
-            '\u{0370}'..='\u{03FF}' |
-            // Cyrillic
-            '\u{0400}'..='\u{04FF}' |
-            // Georgian
-            '\u{10A0}'..='\u{10FF}' |
-            // CJK (treated as LTR in horizontal layout)
-            '\u{4E00}'..='\u{9FFF}' |
-            '\u{3040}'..='\u{309F}' | // Hiragana
-            '\u{30A0}'..='\u{30FF}'   // Katakana
-        )
-    }
-
-    /// Text layout result containing computed metrics.
-    #[derive(Debug, Clone)]
-    pub struct TextLayoutResult {
-        /// Total width of the laid out text.
-        pub width: f32,
-        /// Total height of the laid out text.
-        pub height: f32,
-        /// Number of lines after layout.
-        pub line_count: usize,
-        /// Width of the longest line.
-        pub max_line_width: f32,
-        /// Distance to alphabetic baseline from top.
-        pub alphabetic_baseline: f32,
-    }
-
-    impl TextLayoutResult {
-        /// Returns the size as a Size struct.
-        #[inline]
-        #[must_use]
-        pub fn size(&self) -> Size<Pixels> {
-            Size::new(px(self.width), px(self.height))
-        }
-    }
-
-    /// Extended line information including directionality.
-    #[derive(Debug, Clone, PartialEq)]
-    pub struct LineInfo {
-        /// Line number (0-indexed).
-        pub line_number: usize,
-        /// Whether this line is rendered right-to-left.
-        pub is_rtl: bool,
-        /// Width of the line in pixels.
-        pub width: f32,
-        /// Height of the line in pixels.
-        pub height: f32,
-        /// Top position of the line.
-        pub top: f32,
-        /// Start text index for this line.
-        pub start_index: usize,
-        /// End text index for this line (exclusive).
-        pub end_index: usize,
-    }
-
-    impl LineInfo {
-        /// Returns the text direction for this line.
-        #[inline]
-        #[must_use]
-        pub fn direction(&self) -> TextDirection {
-            if self.is_rtl {
-                TextDirection::Rtl
-            } else {
-                TextDirection::Ltr
-            }
-        }
-
-        /// Returns the bottom position of the line.
-        #[inline]
-        #[must_use]
-        pub fn bottom(&self) -> f32 {
-            self.top + self.height
-        }
-
-        /// Returns the length of text in this line.
-        #[inline]
-        #[must_use]
-        pub fn len(&self) -> usize {
-            self.end_index.saturating_sub(self.start_index)
-        }
-
-        /// Returns true if the line is empty.
-        #[inline]
-        #[must_use]
-        pub fn is_empty(&self) -> bool {
-            self.len() == 0
-        }
-    }
-
-    /// A laid out text buffer with cursor and hit testing support.
-    ///
-    /// This wraps a cosmic-text Buffer and provides methods for:
-    /// - Cursor positioning (offset → screen position)
-    /// - Hit testing (screen position → text offset)
-    /// - Line metrics
-    /// - Selection boxes
-    pub struct TextLayout {
-        /// The underlying cosmic-text buffer.
-        buffer: Buffer,
-        /// Font size used for layout.
-        font_size: f32,
-        /// Line height used for layout.
-        line_height: f32,
-        /// Text direction.
-        direction: TextDirection,
-    }
-
-    impl TextLayout {
-        /// Creates a new text layout.
-        pub fn new(
-            text: &str,
-            style: Option<&TextStyle>,
-            font_size: f32,
-            max_width: Option<f32>,
-            line_height: Option<f32>,
-            direction: TextDirection,
-        ) -> Self {
-            let mut font_system = font_system().lock();
-
-            let line_height = line_height.unwrap_or(font_size * 1.2);
-            let metrics = Metrics::new(font_size, line_height);
-
-            let mut buffer = Buffer::new(&mut font_system, metrics);
-            buffer.set_size(&mut font_system, max_width, None);
-
-            let attrs = style_to_attrs(style);
-            buffer.set_text(&mut font_system, text, attrs, Shaping::Advanced);
-            buffer.shape_until_scroll(&mut font_system, false);
-
-            Self {
-                buffer,
-                font_size,
-                line_height,
-                direction,
-            }
-        }
-
-        /// Returns the computed metrics for this layout.
-        pub fn metrics(&self) -> TextLayoutResult {
-            let mut total_height = 0.0f32;
-            let mut max_line_width = 0.0f32;
-            let mut line_count = 0usize;
-            let mut first_baseline = 0.0f32;
-
-            for run in self.buffer.layout_runs() {
-                line_count += 1;
-                max_line_width = max_line_width.max(run.line_w);
-                total_height = total_height.max(run.line_top + run.line_height);
-
-                if line_count == 1 {
-                    first_baseline = run.line_top + run.line_height * 0.8;
-                }
-            }
-
-            if line_count == 0 {
-                line_count = 1;
-                total_height = self.line_height;
-                first_baseline = self.line_height * 0.8;
-            }
-
-            TextLayoutResult {
-                width: max_line_width,
-                height: total_height,
-                line_count,
-                max_line_width,
-                alphabetic_baseline: first_baseline,
-            }
-        }
-
-        /// Returns the screen offset for a caret at the given text position.
-        ///
-        /// This is used for drawing the text cursor.
-        pub fn get_offset_for_caret(&mut self, position: TextPosition) -> Offset<Pixels> {
-            let mut font_system = font_system().lock();
-
-            // Create cursor at position
-            let cursor = Cursor::new(0, position.offset);
-
-            // Get cursor position from buffer
-            if let Some(layout_cursor) = self.buffer.layout_cursor(&mut font_system, cursor) {
-                // layout_cursor gives us glyph index, we need to find actual x position
-                let mut x = 0.0f32;
-                let mut y = 0.0f32;
-
-                for run in self.buffer.layout_runs() {
-                    if run.line_i == layout_cursor.line {
-                        y = run.line_top;
-
-                        // Find x position by iterating glyphs
-                        for glyph in run.glyphs {
-                            if glyph.start <= position.offset && position.offset <= glyph.end {
-                                // Interpolate within glyph if needed
-                                let glyph_progress = if glyph.end > glyph.start {
-                                    (position.offset - glyph.start) as f32
-                                        / (glyph.end - glyph.start) as f32
-                                } else {
-                                    0.0
-                                };
-                                x = glyph.x + glyph.w * glyph_progress;
-                                break;
-                            }
-                            x = glyph.x + glyph.w;
-                        }
-                        break;
-                    }
-                }
-
-                Offset::new(Pixels(x), Pixels(y))
-            } else {
-                // Fallback for empty buffer or invalid position
-                Offset::ZERO
-            }
-        }
-
-        /// Returns the text position for a screen offset.
-        ///
-        /// This is used for hit testing (e.g., mouse clicks).
-        pub fn get_position_for_offset(&self, offset: Offset<Pixels>) -> TextPosition {
-            let x = offset.dx.0;
-            let y = offset.dy.0;
-
-            // Find the line at y
-            let mut target_line: Option<usize> = None;
-            let mut line_top = 0.0f32;
-
-            for run in self.buffer.layout_runs() {
-                if y >= run.line_top && y < run.line_top + run.line_height {
-                    target_line = Some(run.line_i);
-                    line_top = run.line_top;
-                    break;
-                }
-            }
-
-            // If no line found, use last line or return 0
-            let target_line = if let Some(l) = target_line {
-                l
-            } else {
-                // Check if below all lines
-                let mut last_line = 0;
-                for run in self.buffer.layout_runs() {
-                    last_line = run.line_i;
-                }
-                if y >= line_top {
-                    last_line
-                } else {
-                    return TextPosition::upstream(0);
-                }
-            };
-
-            // Find glyph at x in target line
-            for run in self.buffer.layout_runs() {
-                if run.line_i == target_line {
-                    let mut last_offset = run.glyphs.first().map_or(0, |g| g.start);
-
-                    for glyph in run.glyphs {
-                        let glyph_center = glyph.x + glyph.w / 2.0;
-
-                        if x < glyph_center {
-                            // Cursor is before center of this glyph
-                            return TextPosition::new(glyph.start, TextAffinity::Downstream);
-                        }
-
-                        last_offset = glyph.end;
-                    }
-
-                    // Past all glyphs
-                    return TextPosition::new(last_offset, TextAffinity::Upstream);
-                }
-            }
-
-            TextPosition::upstream(0)
-        }
-
-        /// Returns line metrics for all lines in the layout.
-        pub fn get_line_metrics(&self) -> Vec<LineMetrics> {
-            let mut metrics = Vec::new();
-
-            for (line_number, run) in self.buffer.layout_runs().enumerate() {
-                let ascent = self.font_size * 0.8;
-                let descent = self.font_size * 0.2;
-
-                // Calculate text indices for this line
-                let start_index = run.glyphs.first().map_or(0, |g| g.start);
-                let end_index = run.glyphs.last().map_or(start_index, |g| g.end);
-
-                metrics.push(LineMetrics::new(
-                    true, // hard_break - assume true for now
-                    ascent as f64,
-                    descent as f64,
-                    ascent as f64,                       // unscaled_ascent
-                    run.line_height as f64,              // height
-                    run.line_w as f64,                   // width
-                    0.0,                                 // left
-                    run.line_top as f64 + ascent as f64, // baseline
-                    line_number,
-                    start_index,
-                    end_index,
-                    end_index, // end_excluding_whitespace (simplified)
-                    end_index, // end_including_newline (simplified)
-                ));
-            }
-
-            // Ensure at least one line
-            if metrics.is_empty() {
-                metrics.push(LineMetrics::new(
-                    true,
-                    (self.font_size * 0.8) as f64,
-                    (self.font_size * 0.2) as f64,
-                    (self.font_size * 0.8) as f64,
-                    self.line_height as f64,
-                    0.0,
-                    0.0,
-                    (self.font_size * 0.8) as f64,
-                    0,
-                    0,
-                    0,
-                    0,
-                    0,
-                ));
-            }
-
-            metrics
-        }
-
-        /// Returns extended line information including RTL status.
-        ///
-        /// This provides more detailed information than `get_line_metrics`,
-        /// including whether each line is rendered right-to-left.
-        pub fn get_line_info(&self) -> Vec<LineInfo> {
-            let mut info = Vec::new();
-
-            for run in self.buffer.layout_runs() {
-                let start_index = run.glyphs.first().map_or(0, |g| g.start);
-                let end_index = run.glyphs.last().map_or(start_index, |g| g.end);
-
-                info.push(LineInfo {
-                    line_number: run.line_i,
-                    is_rtl: run.rtl,
-                    width: run.line_w,
-                    height: run.line_height,
-                    top: run.line_top,
-                    start_index,
-                    end_index,
-                });
-            }
-
-            // Ensure at least one line
-            if info.is_empty() {
-                info.push(LineInfo {
-                    line_number: 0,
-                    is_rtl: self.direction.is_rtl(),
-                    width: 0.0,
-                    height: self.line_height,
-                    top: 0.0,
-                    start_index: 0,
-                    end_index: 0,
-                });
-            }
-
-            info
-        }
-
-        /// Returns true if any line in the layout is RTL.
-        pub fn has_rtl_content(&self) -> bool {
-            self.buffer.layout_runs().any(|run| run.rtl)
-        }
-
-        /// Returns true if the layout contains bidirectional text.
-        ///
-        /// This is true when some lines are LTR and others are RTL,
-        /// or when a single line contains mixed directional content.
-        pub fn is_bidirectional(&self) -> bool {
-            let mut has_ltr = false;
-            let mut has_rtl = false;
-
-            for run in self.buffer.layout_runs() {
-                if run.rtl {
-                    has_rtl = true;
-                } else {
-                    has_ltr = true;
-                }
-
-                if has_ltr && has_rtl {
-                    return true;
-                }
-            }
-
-            false
-        }
-
-        /// Returns bounding boxes for the given text range.
-        ///
-        /// Used for rendering text selection highlights.
-        #[allow(clippy::needless_pass_by_value)] // TextRange is small (2 usize), pass by value is idiomatic
-        pub fn get_boxes_for_range(&self, range: TextRange) -> Vec<TextBox> {
-            let mut boxes = Vec::new();
-
-            for run in self.buffer.layout_runs() {
-                let mut line_start_x: Option<f32> = None;
-                let mut line_end_x = 0.0f32;
-
-                for glyph in run.glyphs {
-                    // Check if glyph overlaps with range
-                    if glyph.end > range.start && glyph.start < range.end {
-                        if line_start_x.is_none() {
-                            // Adjust start x if range starts mid-glyph
-                            let start_offset = if glyph.start < range.start {
-                                let progress = (range.start - glyph.start) as f32
-                                    / (glyph.end - glyph.start) as f32;
-                                glyph.x + glyph.w * progress
-                            } else {
-                                glyph.x
-                            };
-                            line_start_x = Some(start_offset);
-                        }
-
-                        // Adjust end x if range ends mid-glyph
-                        line_end_x = if glyph.end > range.end {
-                            let progress =
-                                (range.end - glyph.start) as f32 / (glyph.end - glyph.start) as f32;
-                            glyph.x + glyph.w * progress
-                        } else {
-                            glyph.x + glyph.w
-                        };
-                    }
-                }
-
-                if let Some(start_x) = line_start_x {
-                    let rect = Rect::from_ltrb(
-                        Pixels(start_x),
-                        Pixels(run.line_top),
-                        Pixels(line_end_x),
-                        Pixels(run.line_top + run.line_height),
-                    );
-                    boxes.push(TextBox::new(rect, self.direction));
-                }
-            }
-
-            boxes
-        }
-
-        /// Returns the word boundary at the given text position.
-        ///
-        /// Used for double-click word selection.
-        pub fn get_word_boundary(&self, position: TextPosition) -> TextRange {
-            // Get the full text by reconstructing from glyphs
-            // For now, use a simple word boundary algorithm
-            let mut char_positions: Vec<usize> = Vec::new();
-
-            for run in self.buffer.layout_runs() {
-                for glyph in run.glyphs {
-                    // Each glyph represents a character range
-                    // This is simplified - real implementation would need actual text
-                    char_positions.push(glyph.start);
-                }
-            }
-
-            // Simple word boundary: find whitespace boundaries
-            let offset = position.offset;
-
-            // Find start of word (scan backwards for whitespace or start)
-            let mut word_start = offset;
-            for i in (0..offset).rev() {
-                if char_positions.contains(&i) {
-                    // Check if this is a word boundary
-                    // Simplified: assume no whitespace info available
-                    word_start = i;
-                } else {
-                    break;
-                }
-            }
-
-            // Find end of word (scan forwards for whitespace or end)
-            let mut word_end = offset;
-            let max_pos = char_positions.last().copied().unwrap_or(0);
-            for i in offset..=max_pos {
-                if char_positions.contains(&i) {
-                    word_end = i + 1;
-                } else {
-                    break;
-                }
-            }
-
-            TextRange::new(word_start, word_end)
-        }
-    }
-
-    impl std::fmt::Debug for TextLayout {
-        fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-            f.debug_struct("TextLayout")
-                .field("font_size", &self.font_size)
-                .field("line_height", &self.line_height)
-                .field("direction", &self.direction)
-                .finish_non_exhaustive()
-        }
-    }
-
-    /// Converts FLUI TextStyle to cosmic-text Attrs.
-    fn style_to_attrs(style: Option<&TextStyle>) -> Attrs<'static> {
-        let mut attrs = Attrs::new();
-
-        if let Some(style) = style {
-            // Font family
-            if let Some(ref family) = style.font_family {
-                // For now, use generic families based on name
-                attrs = attrs.family(match family.as_str() {
-                    "serif" | "Serif" => Family::Serif,
-                    "sans-serif" | "SansSerif" | "sans" => Family::SansSerif,
-                    "monospace" | "Monospace" | "mono" => Family::Monospace,
-                    "cursive" | "Cursive" => Family::Cursive,
-                    "fantasy" | "Fantasy" => Family::Fantasy,
-                    _ => Family::SansSerif, // Default
-                });
-            }
-
-            // Font weight
-            if let Some(weight) = style.font_weight {
-                let cosmic_weight = match weight {
-                    FontWeight::W100 => Weight::THIN,
-                    FontWeight::W200 => Weight::EXTRA_LIGHT,
-                    FontWeight::W300 => Weight::LIGHT,
-                    FontWeight::W400 => Weight::NORMAL,
-                    FontWeight::W500 => Weight::MEDIUM,
-                    FontWeight::W600 => Weight::SEMIBOLD,
-                    FontWeight::W700 => Weight::BOLD,
-                    FontWeight::W800 => Weight::EXTRA_BOLD,
-                    FontWeight::W900 => Weight::BLACK,
-                };
-                attrs = attrs.weight(cosmic_weight);
-            }
-
-            // Font style (italic)
-            if let Some(font_style) = style.font_style {
-                let cosmic_style = match font_style {
-                    FontStyle::Normal => Style::Normal,
-                    FontStyle::Italic => Style::Italic,
-                };
-                attrs = attrs.style(cosmic_style);
-            }
-        }
-
-        attrs
-    }
-
-    /// Measures text and returns layout metrics.
-    ///
-    /// # Arguments
-    ///
-    /// * `text` - The text to measure
-    /// * `style` - Optional text style
-    /// * `font_size` - Font size in pixels
-    /// * `max_width` - Maximum width constraint (None for unlimited)
-    /// * `line_height` - Line height in pixels (if None, uses font_size * 1.2)
-    ///
-    /// # Returns
-    ///
-    /// Layout result with computed metrics.
-    pub fn measure_text(
-        text: &str,
-        style: Option<&TextStyle>,
-        font_size: f32,
-        max_width: Option<f32>,
-        line_height: Option<f32>,
-    ) -> TextLayoutResult {
-        let mut font_system = font_system().lock();
-
-        // Create metrics
-        let line_height = line_height.unwrap_or(font_size * 1.2);
-        let metrics = Metrics::new(font_size, line_height);
-
-        // Create buffer
-        let mut buffer = Buffer::new(&mut font_system, metrics);
-
-        // Set size constraint
-        buffer.set_size(&mut font_system, max_width, None);
-
-        // Set text with attributes
-        let attrs = style_to_attrs(style);
-        buffer.set_text(&mut font_system, text, attrs, Shaping::Advanced);
-
-        // Shape the text
-        buffer.shape_until_scroll(&mut font_system, false);
-
-        // Compute metrics from layout runs
-        let mut total_height = 0.0f32;
-        let mut max_line_width = 0.0f32;
-        let mut line_count = 0usize;
-        let mut first_baseline = 0.0f32;
-
-        for run in buffer.layout_runs() {
-            line_count += 1;
-            max_line_width = max_line_width.max(run.line_w);
-            total_height = total_height.max(run.line_top + run.line_height);
-
-            // First line baseline
-            if line_count == 1 {
-                // Approximate alphabetic baseline as ~80% of line height from top
-                first_baseline = run.line_top + run.line_height * 0.8;
-            }
-        }
-
-        // Handle empty text
-        if line_count == 0 {
-            line_count = 1;
-            total_height = line_height;
-            first_baseline = line_height * 0.8;
-        }
-
-        TextLayoutResult {
-            width: max_line_width,
-            height: total_height,
-            line_count,
-            max_line_width,
-            alphabetic_baseline: first_baseline,
-        }
-    }
-
-    /// Measures text with rich spans (InlineSpan).
-    ///
-    /// For now, this extracts plain text and measures it.
-    /// In the future, we can support per-span styling.
-    pub fn measure_inline_span(
-        span: &flui_types::typography::InlineSpan,
-        font_size: f32,
-        max_width: Option<f32>,
-        scale_factor: f32,
-    ) -> TextLayoutResult {
-        let plain_text = span.to_plain_text();
-        let style = span.style();
-        let scaled_font_size = font_size * scale_factor;
-
-        measure_text(&plain_text, style, scaled_font_size, max_width, None)
-    }
-
-    #[cfg(test)]
-    mod tests {
-        use super::*;
-
-        #[test]
-        fn test_measure_simple_text() {
-            let result = measure_text("Hello, World!", None, 14.0, None, None);
-
-            assert!(result.width > 0.0);
-            assert!(result.height > 0.0);
-            assert_eq!(result.line_count, 1);
-        }
-
-        #[test]
-        fn test_measure_multiline_text() {
-            let result = measure_text("Line 1\nLine 2\nLine 3", None, 14.0, None, None);
-
-            assert!(result.height > 0.0);
-            assert_eq!(result.line_count, 3);
-        }
-
-        #[test]
-        fn test_measure_with_width_constraint() {
-            // Long text that should wrap
-            let text = "This is a very long line of text that should wrap when constrained";
-
-            // Without constraint
-            let unconstrained = measure_text(text, None, 14.0, None, None);
-
-            // With narrow constraint
-            let constrained = measure_text(text, None, 14.0, Some(100.0), None);
-
-            // Constrained should have more lines
-            assert!(constrained.line_count >= unconstrained.line_count);
-            // Constrained width should be less
-            assert!(constrained.max_line_width <= 100.0 + 1.0); // Allow small overflow
-        }
-
-        #[test]
-        fn test_measure_empty_text() {
-            let result = measure_text("", None, 14.0, None, None);
-
-            assert_eq!(result.line_count, 1);
-            assert!(result.height > 0.0); // Should still have height
-        }
-
-        #[test]
-        fn test_text_layout_creation() {
-            let layout =
-                TextLayout::new("Hello, World!", None, 14.0, None, None, TextDirection::Ltr);
-
-            let metrics = layout.metrics();
-            assert!(metrics.width > 0.0);
-            assert!(metrics.height > 0.0);
-            assert_eq!(metrics.line_count, 1);
-        }
-
-        #[test]
-        fn test_text_layout_caret_position() {
-            let mut layout = TextLayout::new("Hello", None, 14.0, None, None, TextDirection::Ltr);
-
-            // Position at start
-            let start_offset = layout.get_offset_for_caret(TextPosition::upstream(0));
-            assert!(start_offset.dx >= 0.0);
-
-            // Position in middle
-            let mid_offset = layout.get_offset_for_caret(TextPosition::upstream(2));
-            assert!(mid_offset.dx > start_offset.dx);
-
-            // Position at end
-            let end_offset = layout.get_offset_for_caret(TextPosition::upstream(5));
-            assert!(end_offset.dx >= mid_offset.dx);
-        }
-
-        #[test]
-        fn test_text_layout_hit_test() {
-            let layout = TextLayout::new("Hello", None, 14.0, None, None, TextDirection::Ltr);
-
-            // Hit test at start
-            let pos = layout.get_position_for_offset(Offset::new(px(0.0), px(5.0)));
-            assert_eq!(pos.offset, 0);
-
-            // Hit test past end should give last position
-            let pos = layout.get_position_for_offset(Offset::new(px(1000.0), px(5.0)));
-            assert!(pos.offset <= 5);
-        }
-
-        #[test]
-        fn test_text_layout_line_metrics() {
-            let layout =
-                TextLayout::new("Line 1\nLine 2", None, 14.0, None, None, TextDirection::Ltr);
-
-            let metrics = layout.get_line_metrics();
-            assert_eq!(metrics.len(), 2);
-
-            // First line
-            assert_eq!(metrics[0].line_number, 0);
-            assert!(metrics[0].width > 0.0);
-
-            // Second line
-            assert_eq!(metrics[1].line_number, 1);
-        }
-
-        #[test]
-        fn test_text_layout_selection_boxes() {
-            let layout =
-                TextLayout::new("Hello, World!", None, 14.0, None, None, TextDirection::Ltr);
-
-            // Get boxes for "ello"
-            let boxes = layout.get_boxes_for_range(TextRange::new(1, 5));
-            assert!(!boxes.is_empty());
-
-            // Box should have positive dimensions
-            let first_box = &boxes[0];
-            assert!(first_box.rect.width() > 0.0);
-            assert!(first_box.rect.height() > 0.0);
-        }
-
-        #[test]
-        fn test_text_layout_word_boundary() {
-            let layout = TextLayout::new("Hello World", None, 14.0, None, None, TextDirection::Ltr);
-
-            // Word boundary at position 2 (in "Hello")
-            let boundary = layout.get_word_boundary(TextPosition::upstream(2));
-            // Should return some range containing position 2
-            assert!(boundary.start <= 2);
-            assert!(boundary.end >= 2);
-        }
-
-        // ===== RTL and Bidirectional Tests =====
-
-        #[test]
-        fn test_detect_text_direction_ltr() {
-            assert_eq!(detect_text_direction("Hello"), Some(TextDirection::Ltr));
-            assert_eq!(detect_text_direction("Привет"), Some(TextDirection::Ltr));
-            assert_eq!(detect_text_direction("日本語"), Some(TextDirection::Ltr));
-        }
-
-        #[test]
-        fn test_detect_text_direction_rtl() {
-            assert_eq!(detect_text_direction("مرحبا"), Some(TextDirection::Rtl));
-            assert_eq!(detect_text_direction("שלום"), Some(TextDirection::Rtl));
-        }
-
-        #[test]
-        fn test_detect_text_direction_neutral() {
-            // Numbers and spaces are neutral
-            assert_eq!(detect_text_direction("123"), None);
-            assert_eq!(detect_text_direction("   "), None);
-            assert_eq!(detect_text_direction("!@#$%"), None);
-        }
-
-        #[test]
-        fn test_detect_text_direction_mixed() {
-            // First strong character wins
-            assert_eq!(detect_text_direction("123 Hello"), Some(TextDirection::Ltr));
-            assert_eq!(detect_text_direction("123 مرحبا"), Some(TextDirection::Rtl));
-        }
-
-        #[test]
-        fn test_line_info_ltr() {
-            let layout = TextLayout::new("Hello World", None, 14.0, None, None, TextDirection::Ltr);
-
-            let info = layout.get_line_info();
-            assert_eq!(info.len(), 1);
-            assert_eq!(info[0].line_number, 0);
-            assert!(!info[0].is_rtl);
-            assert_eq!(info[0].direction(), TextDirection::Ltr);
-        }
-
-        #[test]
-        fn test_line_info_rtl() {
-            let layout =
-                TextLayout::new("مرحبا بالعالم", None, 14.0, None, None, TextDirection::Rtl);
-
-            let info = layout.get_line_info();
-            assert_eq!(info.len(), 1);
-            // cosmic-text handles RTL automatically with Shaping::Advanced
-            // The line should be marked as RTL
-        }
-
-        #[test]
-        fn test_has_rtl_content() {
-            let ltr_layout =
-                TextLayout::new("Hello World", None, 14.0, None, None, TextDirection::Ltr);
-            // Pure ASCII text is not RTL
-            assert!(!ltr_layout.has_rtl_content());
-        }
-
-        #[test]
-        fn test_line_info_methods() {
-            let info = LineInfo {
-                line_number: 0,
-                is_rtl: true,
-                width: 100.0,
-                height: 20.0,
-                top: 0.0,
-                start_index: 0,
-                end_index: 10,
-            };
-
-            assert_eq!(info.direction(), TextDirection::Rtl);
-            assert_eq!(info.bottom(), 20.0);
-            assert_eq!(info.len(), 10);
-            assert!(!info.is_empty());
-        }
-
-        #[test]
-        fn test_line_info_empty() {
-            let info = LineInfo {
-                line_number: 0,
-                is_rtl: false,
-                width: 0.0,
-                height: 20.0,
-                top: 0.0,
-                start_index: 5,
-                end_index: 5,
-            };
-
-            assert!(info.is_empty());
-            assert_eq!(info.len(), 0);
-        }
-    }
-}
-
-#[cfg(feature = "text")]
-pub use inner::*;
-
-/// Fallback implementation when text feature is disabled.
-#[cfg(not(feature = "text"))]
-mod fallback {
-    use flui_types::{
-        geometry::{Offset, Pixels, Rect, Size, px},
-        typography::{
-            LineMetrics, TextAffinity, TextBox, TextDirection, TextPosition, TextRange, TextStyle,
-        },
-    };
-
-    /// Text layout result (stub).
-    #[derive(Debug, Clone)]
-    pub struct TextLayoutResult {
-        pub width: f32,
-        pub height: f32,
-        pub line_count: usize,
-        pub max_line_width: f32,
-        pub alphabetic_baseline: f32,
-    }
-
-    impl TextLayoutResult {
-        #[inline]
-        #[must_use]
-        pub fn size(&self) -> Size<Pixels> {
-            Size::new(px(self.width), px(self.height))
-        }
-    }
-
-    /// Extended line information (stub).
-    #[derive(Debug, Clone, PartialEq)]
-    pub struct LineInfo {
-        pub line_number: usize,
-        pub is_rtl: bool,
-        pub width: f32,
-        pub height: f32,
-        pub top: f32,
-        pub start_index: usize,
-        pub end_index: usize,
-    }
-
-    impl LineInfo {
-        #[inline]
-        #[must_use]
-        pub fn direction(&self) -> TextDirection {
-            if self.is_rtl {
-                TextDirection::Rtl
-            } else {
-                TextDirection::Ltr
-            }
-        }
-
-        #[inline]
-        #[must_use]
-        pub fn bottom(&self) -> f32 {
-            self.top + self.height
-        }
-
-        #[inline]
-        #[must_use]
-        pub fn len(&self) -> usize {
-            self.end_index.saturating_sub(self.start_index)
-        }
-
-        #[inline]
-        #[must_use]
-        pub fn is_empty(&self) -> bool {
-            self.len() == 0
-        }
-    }
-
-    /// Stub text layout for use by `TextPainter` when the "text" feature is disabled.
-    ///
-    /// Provides estimated metrics based on character count rather than real shaping.
-    #[derive(Debug)]
-    pub struct TextLayout {
-        /// Font size used for layout.
-        font_size: f32,
-        /// Line height used for layout.
-        line_height: f32,
-        /// Text direction.
-        direction: TextDirection,
-        /// The plain text stored for hit testing / metrics.
-        text: String,
-        /// Estimated width.
-        width: f32,
-        /// Number of estimated lines.
-        line_count: usize,
-    }
-
-    impl TextLayout {
-        /// Creates a new stub text layout with estimated metrics.
-        pub fn new(
-            text: &str,
-            _style: Option<&TextStyle>,
-            font_size: f32,
-            max_width: Option<f32>,
-            line_height: Option<f32>,
-            direction: TextDirection,
-        ) -> Self {
-            let line_height = line_height.unwrap_or(font_size * 1.2);
-            let char_count = text.chars().count();
-            let estimated_width = char_count as f32 * font_size * 0.5;
-
-            let (width, line_count) = if let Some(max_w) = max_width {
-                if estimated_width > max_w {
-                    let lines = (estimated_width / max_w).ceil() as usize;
-                    (max_w.min(estimated_width), lines.max(1))
-                } else {
-                    (estimated_width, 1)
-                }
-            } else {
-                (estimated_width, 1)
-            };
-
-            Self {
-                font_size,
-                line_height,
-                direction,
-                text: text.to_string(),
-                width,
-                line_count,
-            }
-        }
-
-        /// Returns estimated metrics.
-        pub fn metrics(&self) -> TextLayoutResult {
-            TextLayoutResult {
-                width: self.width,
-                height: self.line_count as f32 * self.line_height,
-                line_count: self.line_count,
-                max_line_width: self.width,
-                alphabetic_baseline: self.line_height * 0.8,
-            }
-        }
-
-        /// Returns the screen offset for a caret at the given text position (estimated).
-        pub fn get_offset_for_caret(&mut self, position: TextPosition) -> Offset<Pixels> {
-            // Estimate x position based on character index and average char width
-            let avg_char_width = self.font_size * 0.5;
-            let x = position.offset as f32 * avg_char_width;
-            Offset::new(Pixels(x), Pixels(0.0))
-        }
-
-        /// Returns the text position for a screen offset (estimated).
-        pub fn get_position_for_offset(&self, offset: Offset<Pixels>) -> TextPosition {
-            let avg_char_width = self.font_size * 0.5;
-            let char_index = if avg_char_width > 0.0 {
-                (offset.dx.0 / avg_char_width).round() as usize
-            } else {
-                0
-            };
-            let max_offset = self.text.len();
-            TextPosition::new(char_index.min(max_offset), TextAffinity::Downstream)
-        }
-
-        /// Returns line metrics for all lines in the layout (estimated).
-        pub fn get_line_metrics(&self) -> Vec<LineMetrics> {
-            let mut metrics = Vec::with_capacity(self.line_count.max(1));
-            let ascent = self.font_size * 0.8;
-            let descent = self.font_size * 0.2;
-
-            for i in 0..self.line_count.max(1) {
-                metrics.push(LineMetrics::new(
-                    true,
-                    ascent as f64,
-                    descent as f64,
-                    ascent as f64,
-                    self.line_height as f64,
-                    self.width as f64,
-                    0.0,
-                    (i as f32 * self.line_height + ascent) as f64,
-                    i,
-                    0,
-                    0,
-                    0,
-                    0,
-                ));
-            }
-
-            if metrics.is_empty() {
-                metrics.push(LineMetrics::new(
-                    true,
-                    ascent as f64,
-                    descent as f64,
-                    ascent as f64,
-                    self.line_height as f64,
-                    0.0,
-                    0.0,
-                    ascent as f64,
-                    0,
-                    0,
-                    0,
-                    0,
-                    0,
-                ));
-            }
-
-            metrics
-        }
-
-        /// Returns bounding boxes for the given text range (estimated).
-        #[allow(clippy::needless_pass_by_value)]
-        pub fn get_boxes_for_range(&self, range: TextRange) -> Vec<TextBox> {
-            let avg_char_width = self.font_size * 0.5;
-            let start_x = range.start as f32 * avg_char_width;
-            let end_x = range.end as f32 * avg_char_width;
-
-            let rect = Rect::from_ltrb(
-                Pixels(start_x),
-                Pixels(0.0),
-                Pixels(end_x),
-                Pixels(self.line_height),
-            );
-
-            vec![TextBox::new(rect, self.direction)]
-        }
-
-        /// Returns the word boundary at the given text position (estimated).
-        pub fn get_word_boundary(&self, position: TextPosition) -> TextRange {
-            // Simple word boundary: find whitespace boundaries
-            let bytes = self.text.as_bytes();
-            let offset = position.offset.min(bytes.len());
-
-            let mut start = offset;
-            while start > 0 && bytes.get(start - 1).map_or(false, |&b| b != b' ') {
-                start -= 1;
-            }
-
-            let mut end = offset;
-            while end < bytes.len() && bytes.get(end).map_or(false, |&b| b != b' ') {
-                end += 1;
-            }
-
-            TextRange::new(start, end)
-        }
-    }
-
-    /// Stub direction detection.
-    pub fn detect_text_direction(_text: &str) -> Option<TextDirection> {
-        Some(TextDirection::Ltr)
-    }
-
-    /// Stub measurement (estimates based on character count).
-    pub fn measure_text(
-        text: &str,
-        _style: Option<&TextStyle>,
-        font_size: f32,
-        max_width: Option<f32>,
-        _line_height: Option<f32>,
-    ) -> TextLayoutResult {
-        let char_count = text.chars().count();
-        let estimated_width = char_count as f32 * font_size * 0.5;
-        let line_height = font_size * 1.2;
-
-        let (width, line_count) = if let Some(max_w) = max_width {
-            if estimated_width > max_w {
-                let lines = (estimated_width / max_w).ceil() as usize;
-                (max_w.min(estimated_width), lines.max(1))
-            } else {
-                (estimated_width, 1)
-            }
-        } else {
-            (estimated_width, 1)
-        };
-
-        TextLayoutResult {
-            width,
-            height: line_count as f32 * line_height,
-            line_count,
-            max_line_width: width,
-            alphabetic_baseline: line_height * 0.8,
-        }
-    }
-
-    /// Stub measurement for InlineSpan.
-    pub fn measure_inline_span(
-        span: &flui_types::typography::InlineSpan,
-        font_size: f32,
-        max_width: Option<f32>,
-        scale_factor: f32,
-    ) -> TextLayoutResult {
-        let plain_text = span.to_plain_text();
-        let scaled_font_size = font_size * scale_factor;
-        measure_text(&plain_text, span.style(), scaled_font_size, max_width, None)
-    }
-}
-
-#[cfg(not(feature = "text"))]
-pub use fallback::*;
